refactor(runner): extract shellSplit to shell-split.ts
