+++ src/hitl.ts
@@ -32,7 +32,7 @@ import { execQuiet } from "./exec.ts";
 import { DONE_LABEL, IN_PROGRESS_LABEL, STUCK_LABEL } from "./labels.ts";
 import { prepareWorktree, type SetupSandboxConfig } from "./worktree/index.ts";
 import { resolveMainRepo, ensureRepoHasCommit } from "./worktree/management.ts";
-import { shellSplit } from "./runner.ts";
+import { shellSplit } from "./shell-split.ts";
 import { formatFileRef } from "./prompt.ts";
 
 // ---------------------------------------------------------------------------
